requirements.txt
discord
python-dotenv
paramiko